requirements.txt
# Core
python-dotenv==1.0.0
numpy==1.26.4
scipy==1.13.0
pydantic==2.7.0

# Sentence transformers (Layer 3 — runs locally, no API)
sentence-transformers==3.0.1
torch==2.3.0
transformers==4.41.0

# Neural ODE (Layer 5 — Chen et al. NeurIPS 2018). Optional; fallback Euler if missing.
torchdiffeq>=0.2.0

# LLM APIs (use whichever you have credits for)
google-generativeai==0.7.0
openai==1.30.0

# Terminal UI
rich==13.7.1

# Datasets (for download_datasets.py and training on InjecAgent / deepset / LLMail)
datasets>=2.14.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7

# Web dashboard (stretch goal)
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
